Simplify things a bit and remove the extraneous runtime transformer for 1HE.
